src/at/rylicx/adventureguild/constant/Archetypes.java
package at.rylicx.adventureguild.constant;

public enum Archetypes {
    NONE("Undefined"),
    BARBARIAN("Barbarian"),
    BARD("Bard"),
    CLERIC("Cleric"),
    DRUID("Druid"),
    FIGHTER("Fighter"),
    MONK("Monk"),
    PALADIN("Paladin"),
    RANGER("Ranger"),
    ROGUE("Rogue"),
    SORCERER("Sorcerer"),
    WARLOCK("Warlock"),
    Wizard("Wizard");

    private final String displayName;

    Archetypes(String displayName) {
        this.displayName = displayName;
    }

    public String getDisplayName() {
        return displayName;
    }
}